refactor: split internal/jam/message.go by record type
726 lines covering every JAM record operation, split along the record each
group touches:

  message.go                 276  ReadMessage/WriteMessage/DeleteMessage/Scan
  message_header.go          217  header read + write, including subfields
  message_header_update.go    99  in-place header rewrite
  subfields.go                56  subfield construction, origin-line parsing
  message_text.go             52  .jdt text read and append
  binary.go                   35  little-endian read/write helpers
  cp437.go                    35  CP437 to Unicode conversion

Pure move: no declaration body changed. Verified with a go/parser pairwise
diff of every top-level declaration, printed and compared — 18 old, 18 new,
0 missing, 0 mismatched, 0 extra. No test file touched.
