+++ internal/jam/message.go
@@ -1,382 +1,14 @@
 package jam
 
 import (
-	"encoding/binary"
 	"fmt"
-	"io"
 	"os"
 	"strings"
 	"time"
 )
 
-func readBinaryLE(r io.Reader, data interface{}, label string) error {
-	if err := binary.Read(r, binary.LittleEndian, data); err != nil {
-		return fmt.Errorf("jam: read %s: %w", label, err)
-	}
-	return nil
-}
-
-func writeBinaryLE(w io.Writer, data interface{}, label string) error {
-	if err := binary.Write(w, binary.LittleEndian, data); err != nil {
-		return fmt.Errorf("jam: write %s: %w", label, err)
-	}
-	return nil
-}
-
-func writeAll(w io.Writer, buffer []byte, label string) error {
-	n, err := w.Write(buffer)
-	if err != nil {
-		return fmt.Errorf("jam: write %s: %w", label, err)
-	}
-	if n != len(buffer) {
-		return fmt.Errorf("jam: short write %s: wrote %d of %d bytes", label, n, len(buffer))
-	}
-	return nil
-}
-
-// ReadMessageHeader reads a message header for the given 1-based message number.
-func (b *Base) ReadMessageHeader(msgNum int) (*MessageHeader, error) {
-	b.mu.RLock()
-	defer b.mu.RUnlock()
-	return b.readMessageHeaderLocked(msgNum)
-}
-
-func (b *Base) readMessageHeaderLocked(msgNum int) (*MessageHeader, error) {
-	if !b.isOpen {
-		return nil, ErrBaseNotOpen
-	}
-
-	idx, err := b.readIndexRecordLocked(msgNum)
-	if err != nil {
-		return nil, err
-	}
-
-	if _, err := b.jhrFile.Seek(int64(idx.HdrOffset), 0); err != nil {
-		return nil, fmt.Errorf("jam: seek failed on .jhr: %w", err)
-	}
-
-	hdr := &MessageHeader{}
-	if err := readBinaryLE(b.jhrFile, &hdr.Signature, "header signature"); err != nil {
-		return nil, err
-	}
-	if err := readBinaryLE(b.jhrFile, &hdr.Revision, "header revision"); err != nil {
-		return nil, err
-	}
-	if err := readBinaryLE(b.jhrFile, &hdr.ReservedWord, "header reserved word"); err != nil {
-		return nil, err
-	}
-	if err := readBinaryLE(b.jhrFile, &hdr.SubfieldLen, "header subfield length"); err != nil {
-		return nil, err
-	}
-	if err := readBinaryLE(b.jhrFile, &hdr.TimesRead, "header times read"); err != nil {
-		return nil, err
-	}
-	if err := readBinaryLE(b.jhrFile, &hdr.MSGIDcrc, "header MSGID crc"); err != nil {
-		return nil, err
-	}
-	if err := readBinaryLE(b.jhrFile, &hdr.REPLYcrc, "header REPLY crc"); err != nil {
-		return nil, err
-	}
-	if err := readBinaryLE(b.jhrFile, &hdr.ReplyTo, "header reply to"); err != nil {
-		return nil, err
-	}
-	if err := readBinaryLE(b.jhrFile, &hdr.Reply1st, "header reply first"); err != nil {
-		return nil, err
-	}
-	if err := readBinaryLE(b.jhrFile, &hdr.ReplyNext, "header reply next"); err != nil {
-		return nil, err
-	}
-	if err := readBinaryLE(b.jhrFile, &hdr.DateWritten, "header date written"); err != nil {
-		return nil, err
-	}
-	if err := readBinaryLE(b.jhrFile, &hdr.DateReceived, "header date received"); err != nil {
-		return nil, err
-	}
-	if err := readBinaryLE(b.jhrFile, &hdr.DateProcessed, "header date processed"); err != nil {
-		return nil, err
-	}
-	if err := readBinaryLE(b.jhrFile, &hdr.MessageNumber, "header message number"); err != nil {
-		return nil, err
-	}
-	if err := readBinaryLE(b.jhrFile, &hdr.Attribute, "header attribute"); err != nil {
-		return nil, err
-	}
-	if err := readBinaryLE(b.jhrFile, &hdr.Attribute2, "header attribute2"); err != nil {
-		return nil, err
-	}
-	if err := readBinaryLE(b.jhrFile, &hdr.Offset, "header text offset"); err != nil {
-		return nil, err
-	}
-	if err := readBinaryLE(b.jhrFile, &hdr.TxtLen, "header text length"); err != nil {
-		return nil, err
-	}
-	if err := readBinaryLE(b.jhrFile, &hdr.PasswordCRC, "header password crc"); err != nil {
-		return nil, err
-	}
-	if err := readBinaryLE(b.jhrFile, &hdr.Cost, "header cost"); err != nil {
-		return nil, err
-	}
-
-	if string(hdr.Signature[:]) != Signature {
-		return nil, ErrInvalidSignature
-	}
-
-	// Read subfields
-	bytesRead := uint32(0)
-	for bytesRead < hdr.SubfieldLen {
-		sf := Subfield{}
-		if err := readBinaryLE(b.jhrFile, &sf.LoID, "subfield loID"); err != nil {
-			return nil, err
-		}
-		if err := readBinaryLE(b.jhrFile, &sf.HiID, "subfield hiID"); err != nil {
-			return nil, err
-		}
-		if err := readBinaryLE(b.jhrFile, &sf.DatLen, "subfield data length"); err != nil {
-			return nil, err
-		}
-		sf.Buffer = make([]byte, sf.DatLen)
-		if _, err := io.ReadFull(b.jhrFile, sf.Buffer); err != nil {
-			return nil, fmt.Errorf("jam: read subfield buffer: %w", err)
-		}
-		hdr.Subfields = append(hdr.Subfields, sf)
-		bytesRead += SubfieldHdrSize + sf.DatLen
-	}
-
-	return hdr, nil
-}
-
-// WriteMessageHeader writes a message header to the .jhr file and returns
-// the byte offset where it was written.
-func (b *Base) writeMessageHeader(hdr *MessageHeader) (uint32, error) {
-	if !b.isOpen {
-		return 0, ErrBaseNotOpen
-	}
-
-	pos, err := b.jhrFile.Seek(0, io.SeekEnd)
-	if err != nil {
-		return 0, fmt.Errorf("jam: seek failed on .jhr: %w", err)
-	}
-	if pos == 0 {
-		pos = HeaderSize
-		if _, err := b.jhrFile.Seek(pos, 0); err != nil {
-			return 0, fmt.Errorf("jam: seek failed on .jhr: %w", err)
-		}
-	}
-
-	if err := writeBinaryLE(b.jhrFile, hdr.Signature, "header signature"); err != nil {
-		return 0, err
-	}
-	if err := writeBinaryLE(b.jhrFile, hdr.Revision, "header revision"); err != nil {
-		return 0, err
-	}
-	if err := writeBinaryLE(b.jhrFile, hdr.ReservedWord, "header reserved word"); err != nil {
-		return 0, err
-	}
-	if err := writeBinaryLE(b.jhrFile, hdr.SubfieldLen, "header subfield length"); err != nil {
-		return 0, err
-	}
-	if err := writeBinaryLE(b.jhrFile, hdr.TimesRead, "header times read"); err != nil {
-		return 0, err
-	}
-	if err := writeBinaryLE(b.jhrFile, hdr.MSGIDcrc, "header MSGID crc"); err != nil {
-		return 0, err
-	}
-	if err := writeBinaryLE(b.jhrFile, hdr.REPLYcrc, "header REPLY crc"); err != nil {
-		return 0, err
-	}
-	if err := writeBinaryLE(b.jhrFile, hdr.ReplyTo, "header reply to"); err != nil {
-		return 0, err
-	}
-	if err := writeBinaryLE(b.jhrFile, hdr.Reply1st, "header reply first"); err != nil {
-		return 0, err
-	}
-	if err := writeBinaryLE(b.jhrFile, hdr.ReplyNext, "header reply next"); err != nil {
-		return 0, err
-	}
-	if err := writeBinaryLE(b.jhrFile, hdr.DateWritten, "header date written"); err != nil {
-		return 0, err
-	}
-	if err := writeBinaryLE(b.jhrFile, hdr.DateReceived, "header date received"); err != nil {
-		return 0, err
-	}
-	if err := writeBinaryLE(b.jhrFile, hdr.DateProcessed, "header date processed"); err != nil {
-		return 0, err
-	}
-	if err := writeBinaryLE(b.jhrFile, hdr.MessageNumber, "header message number"); err != nil {
-		return 0, err
-	}
-	if err := writeBinaryLE(b.jhrFile, hdr.Attribute, "header attribute"); err != nil {
-		return 0, err
-	}
-	if err := writeBinaryLE(b.jhrFile, hdr.Attribute2, "header attribute2"); err != nil {
-		return 0, err
-	}
-	if err := writeBinaryLE(b.jhrFile, hdr.Offset, "header text offset"); err != nil {
-		return 0, err
-	}
-	if err := writeBinaryLE(b.jhrFile, hdr.TxtLen, "header text length"); err != nil {
-		return 0, err
-	}
-	if err := writeBinaryLE(b.jhrFile, hdr.PasswordCRC, "header password crc"); err != nil {
-		return 0, err
-	}
-	if err := writeBinaryLE(b.jhrFile, hdr.Cost, "header cost"); err != nil {
-		return 0, err
-	}
-
-	for _, sf := range hdr.Subfields {
-		if err := writeBinaryLE(b.jhrFile, sf.LoID, "subfield loID"); err != nil {
-			return 0, err
-		}
-		if err := writeBinaryLE(b.jhrFile, sf.HiID, "subfield hiID"); err != nil {
-			return 0, err
-		}
-		if err := writeBinaryLE(b.jhrFile, sf.DatLen, "subfield data length"); err != nil {
-			return 0, err
-		}
-		if err := writeAll(b.jhrFile, sf.Buffer, "subfield buffer"); err != nil {
-			return 0, err
-		}
-	}
-
-	return uint32(pos), nil
-}
-
-// UpdateMessageHeader rewrites an existing message header in place.
-// This is used by the tosser to update DateProcessed after export.
-func (b *Base) UpdateMessageHeader(msgNum int, hdr *MessageHeader) error {
-	return b.withFileLock(func() error {
-		b.mu.Lock()
-		defer b.mu.Unlock()
-		return b.updateMessageHeaderLocked(msgNum, hdr)
-	})
-}
-
-// updateMessageHeaderLocked is the lock-free internal implementation.
-// Caller must hold both the file lock and b.mu.
-func (b *Base) updateMessageHeaderLocked(msgNum int, hdr *MessageHeader) error {
-	if !b.isOpen {
-		return ErrBaseNotOpen
-	}
-
-	idx, err := b.readIndexRecordLocked(msgNum)
-	if err != nil {
-		return err
-	}
-
-	if _, err := b.jhrFile.Seek(int64(idx.HdrOffset), 0); err != nil {
-		return fmt.Errorf("jam: seek failed on .jhr: %w", err)
-	}
-
-	if err := writeBinaryLE(b.jhrFile, hdr.Signature, "header signature"); err != nil {
-		return err
-	}
-	if err := writeBinaryLE(b.jhrFile, hdr.Revision, "header revision"); err != nil {
-		return err
-	}
-	if err := writeBinaryLE(b.jhrFile, hdr.ReservedWord, "header reserved word"); err != nil {
-		return err
-	}
-	if err := writeBinaryLE(b.jhrFile, hdr.SubfieldLen, "header subfield length"); err != nil {
-		return err
-	}
-	if err := writeBinaryLE(b.jhrFile, hdr.TimesRead, "header times read"); err != nil {
-		return err
-	}
-	if err := writeBinaryLE(b.jhrFile, hdr.MSGIDcrc, "header MSGID crc"); err != nil {
-		return err
-	}
-	if err := writeBinaryLE(b.jhrFile, hdr.REPLYcrc, "header REPLY crc"); err != nil {
-		return err
-	}
-	if err := writeBinaryLE(b.jhrFile, hdr.ReplyTo, "header reply to"); err != nil {
-		return err
-	}
-	if err := writeBinaryLE(b.jhrFile, hdr.Reply1st, "header reply first"); err != nil {
-		return err
-	}
-	if err := writeBinaryLE(b.jhrFile, hdr.ReplyNext, "header reply next"); err != nil {
-		return err
-	}
-	if err := writeBinaryLE(b.jhrFile, hdr.DateWritten, "header date written"); err != nil {
-		return err
-	}
-	if err := writeBinaryLE(b.jhrFile, hdr.DateReceived, "header date received"); err != nil {
-		return err
-	}
-	if err := writeBinaryLE(b.jhrFile, hdr.DateProcessed, "header date processed"); err != nil {
-		return err
-	}
-	if err := writeBinaryLE(b.jhrFile, hdr.MessageNumber, "header message number"); err != nil {
-		return err
-	}
-	if err := writeBinaryLE(b.jhrFile, hdr.Attribute, "header attribute"); err != nil {
-		return err
-	}
-	if err := writeBinaryLE(b.jhrFile, hdr.Attribute2, "header attribute2"); err != nil {
-		return err
-	}
-	if err := writeBinaryLE(b.jhrFile, hdr.Offset, "header text offset"); err != nil {
-		return err
-	}
-	if err := writeBinaryLE(b.jhrFile, hdr.TxtLen, "header text length"); err != nil {
-		return err
-	}
-	if err := writeBinaryLE(b.jhrFile, hdr.PasswordCRC, "header password crc"); err != nil {
-		return err
-	}
-	if err := writeBinaryLE(b.jhrFile, hdr.Cost, "header cost"); err != nil {
-		return err
-	}
-
-	// Note: subfields are not rewritten since they don't change
-	// and follow immediately after the fixed header portion.
-
-	return nil
-}
-
-// ReadMessageText reads the raw message text (CP437) for the given header.
-func (b *Base) ReadMessageText(hdr *MessageHeader) (string, error) {
-	b.mu.RLock()
-	defer b.mu.RUnlock()
-	return b.readMessageTextLocked(hdr)
-}
-
-func (b *Base) readMessageTextLocked(hdr *MessageHeader) (string, error) {
-	if !b.isOpen {
-		return "", ErrBaseNotOpen
-	}
-	if hdr.TxtLen == 0 {
-		return "", nil
-	}
-	if _, err := b.jdtFile.Seek(int64(hdr.Offset), 0); err != nil {
-		return "", fmt.Errorf("jam: seek failed on .jdt: %w", err)
-	}
-	buf := make([]byte, hdr.TxtLen)
-	if _, err := io.ReadFull(b.jdtFile, buf); err != nil {
-		return "", fmt.Errorf("jam: failed to read text: %w", err)
-	}
-	return string(buf), nil
-}
-
-// writeMessageText appends text to the .jdt file. LF is converted to CR
-// per the JAM specification. Returns the offset and byte length written.
-func (b *Base) writeMessageText(text string) (uint32, uint32, error) {
-	if !b.isOpen {
-		return 0, 0, ErrBaseNotOpen
-	}
-	text = strings.ReplaceAll(text, "\r\n", "\r") // Handle Windows line endings first
-	text = strings.ReplaceAll(text, "\n", "\r")   // Then convert remaining Unix LF to CR
-	pos, err := b.jdtFile.Seek(0, io.SeekEnd)
-	if err != nil {
-		return 0, 0, fmt.Errorf("jam: seek failed on .jdt: %w", err)
-	}
-	buf := []byte(text)
-	if _, err := b.jdtFile.Write(buf); err != nil {
-		return 0, 0, fmt.Errorf("jam: failed to write text: %w", err)
-	}
-	return uint32(pos), uint32(len(buf)), nil
-}
+// Whole-message operations: the read/write/delete/scan entry points that
+// combine a header, its subfields and its text.
 
 // ReadMessage reads a complete message (header + subfields + text) for
 // the given 1-based message number.
@@ -641,86 +273,4 @@ func (b *Base) ScanMessages(startMsg, maxMessages int) ([]*Message, error) {
 		read++
 	}
 	return messages, scanErr
-}
-
-// buildSubfields assembles the standard subfield list for a message.
-func buildSubfields(msg *Message) []Subfield {
-	var sfs []Subfield
-	if msg.OrigAddr != "" {
-		sfs = append(sfs, CreateSubfield(SfldOAddress, msg.OrigAddr))
-	}
-	if msg.DestAddr != "" {
-		sfs = append(sfs, CreateSubfield(SfldDAddress, msg.DestAddr))
-	}
-	if msg.From != "" {
-		sfs = append(sfs, CreateSubfield(SfldSenderName, msg.From))
-	}
-	if msg.To != "" {
-		sfs = append(sfs, CreateSubfield(SfldReceiverName, msg.To))
-	}
-	if msg.Subject != "" {
-		sfs = append(sfs, CreateSubfield(SfldSubject, msg.Subject))
-	}
-	if msg.MsgID != "" {
-		sfs = append(sfs, CreateSubfield(SfldMsgID, msg.MsgID))
-	}
-	if msg.ReplyID != "" {
-		sfs = append(sfs, CreateSubfield(SfldReplyID, msg.ReplyID))
-	}
-	if msg.PID != "" {
-		sfs = append(sfs, CreateSubfield(SfldPID, msg.PID))
-	}
-	for _, kludge := range msg.Kludges {
-		sfs = append(sfs, CreateSubfield(SfldFTSKludge, kludge))
-	}
-	return sfs
-}
-
-// extractAddressFromOriginLine parses a FidoNet address from the origin line.
-// Format: " * Origin: BBS Name (address)"
-func extractAddressFromOriginLine(text string) string {
-	normalized := strings.ReplaceAll(strings.ReplaceAll(text, "\r\n", "\n"), "\r", "\n")
-	for _, line := range strings.Split(normalized, "\n") {
-		line = strings.TrimSpace(line)
-		if strings.HasPrefix(line, "* Origin:") {
-			start := strings.LastIndex(line, "(")
-			end := strings.LastIndex(line, ")")
-			if start != -1 && end != -1 && end > start {
-				return strings.TrimSpace(line[start+1 : end])
-			}
-		}
-	}
-	return ""
-}
-
-// CP437ToUnicode converts CP437-encoded bytes to a Unicode string.
-func CP437ToUnicode(input []byte) string {
-	cp437 := []rune{
-		0x00C7, 0x00FC, 0x00E9, 0x00E2, 0x00E4, 0x00E0, 0x00E5, 0x00E7,
-		0x00EA, 0x00EB, 0x00E8, 0x00EF, 0x00EE, 0x00EC, 0x00C4, 0x00C5,
-		0x00C9, 0x00E6, 0x00C6, 0x00F4, 0x00F6, 0x00F2, 0x00FB, 0x00F9,
-		0x00FF, 0x00D6, 0x00DC, 0x00A2, 0x00A3, 0x00A5, 0x20A7, 0x0192,
-		0x00E1, 0x00ED, 0x00F3, 0x00FA, 0x00F1, 0x00D1, 0x00AA, 0x00BA,
-		0x00BF, 0x2310, 0x00AC, 0x00BD, 0x00BC, 0x00A1, 0x00AB, 0x00BB,
-		0x2591, 0x2592, 0x2593, 0x2502, 0x2524, 0x2561, 0x2562, 0x2556,
-		0x2555, 0x2563, 0x2551, 0x2557, 0x255D, 0x255C, 0x255B, 0x2510,
-		0x2514, 0x2534, 0x252C, 0x251C, 0x2500, 0x253C, 0x255E, 0x255F,
-		0x255A, 0x2554, 0x2569, 0x2566, 0x2560, 0x2550, 0x256C, 0x2567,
-		0x2568, 0x2564, 0x2565, 0x2559, 0x2558, 0x2552, 0x2553, 0x256B,
-		0x256A, 0x2518, 0x250C, 0x2588, 0x2584, 0x258C, 0x2590, 0x2580,
-		0x03B1, 0x00DF, 0x0393, 0x03C0, 0x03A3, 0x03C3, 0x00B5, 0x03C4,
-		0x03A6, 0x0398, 0x03A9, 0x03B4, 0x221E, 0x03C6, 0x03B5, 0x2229,
-		0x2261, 0x00B1, 0x2265, 0x2264, 0x2320, 0x2321, 0x00F7, 0x2248,
-		0x00B0, 0x2219, 0x00B7, 0x221A, 0x207F, 0x00B2, 0x25A0, 0x00A0,
-	}
-
-	result := make([]rune, 0, len(input))
-	for _, ch := range input {
-		if ch < 0x80 {
-			result = append(result, rune(ch))
-		} else {
-			result = append(result, cp437[ch-0x80])
-		}
-	}
-	return string(result)
 }
